requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
websockets>=12.0
jinja2>=3.1.0
aiofiles>=23.2.0
scikit-learn>=1.4.0
PyNaCl>=1.5.0
PyYAML>=6.0.1

# Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-logging>=0.45b0
opentelemetry-exporter-otlp>=1.24.0
